refactor(back): split up services in MapsModule
This service could get unwieldly with incoming map submission stuff, makes sense to split it up.
