+++ apps/backend/src/app/modules/maps/maps.module.ts
@@ -8,6 +8,10 @@ import { MapLibraryService } from './map-library.service';
 import { SteamModule } from '../steam/steam.module';
 import { RunsModule } from '../runs/runs.module';
 import { RanksModule } from '../ranks/ranks.module';
+import { MapReviewService } from './map-review.service';
+import { MapCreditsService } from './map-credits.service';
+import { MapImageService } from './map-image.service';
+import { MapSubmissionService } from './map-submission.service';
 
 @Module({
   imports: [
@@ -19,7 +23,14 @@ import { RanksModule } from '../ranks/ranks.module';
     RanksModule
   ],
   controllers: [MapsController],
-  providers: [MapsService, MapLibraryService],
-  exports: [MapsService, MapLibraryService]
+  providers: [
+    MapsService,
+    MapLibraryService,
+    MapReviewService,
+    MapCreditsService,
+    MapImageService,
+    MapSubmissionService
+  ],
+  exports: [MapsService, MapLibraryService, MapSubmissionService]
 })
 export class MapsModule {}
