+++ apps/backend/src/app/modules/maps/maps.service.ts
@@ -1,5 +1,4 @@
 import {
-  BadGatewayException,
   BadRequestException,
   ConflictException,
   ForbiddenException,
@@ -9,7 +8,6 @@ import {
 } from '@nestjs/common';
 import {
   Map as MapDB,
-  MapCredit,
   MapTrack,
   MapZone,
   MapZoneTrigger,
@@ -17,36 +15,25 @@ import {
   Prisma
 } from '@prisma/client';
 import { FileStoreCloudService } from '../filestore/file-store-cloud.service';
-import { FileStoreCloudFile } from '../filestore/file-store.interface';
 import { ConfigService } from '@nestjs/config';
-import sharp from 'sharp';
 import { RunsService } from '../runs/runs.service';
 import {
   AdminCtlMapsGetAllQueryDto,
-  CreateMapCreditDto,
   CreateMapDto,
   DtoFactory,
   expandToPrismaIncludes,
-  MapCreditDto,
   MapDto,
-  MapImageDto,
   MapInfoDto,
   MapsCtlGetAllQueryDto,
   MapTrackDto,
   PagedResponseDto,
-  UpdateMapCreditDto,
   UpdateMapDto,
   UpdateMapInfoDto
 } from '@momentum/backend/dto';
-import {
-  ActivityType,
-  MapCreditType,
-  MapStatus,
-  Role
-} from '@momentum/constants';
+import { ActivityType, MapCreditType, MapStatus } from '@momentum/constants';
 import { isEmpty } from 'lodash';
-import { Bitflags } from '@momentum/bitflags';
 import { DbService } from '../database/db.service';
+import { MapImageService } from './map-image.service';
 
 @Injectable()
 export class MapsService {
@@ -54,7 +41,8 @@ export class MapsService {
     private readonly db: DbService,
     private readonly fileCloudService: FileStoreCloudService,
     private readonly config: ConfigService,
-    private readonly runsService: RunsService
+    private readonly runsService: RunsService,
+    private readonly mapImageService: MapImageService
   ) {}
 
   //#region Maps
@@ -532,7 +520,9 @@ export class MapsService {
 
     // Delete all stored map images
     const images = await this.db.mapImage.findMany({ where: { mapID } });
-    await Promise.all(images.map((img) => this.deleteStoredMapImage(img.id)));
+    await Promise.all(
+      images.map((img) => this.mapImageService.deleteStoredMapImage(img.id))
+    );
 
     // Delete all run files
     await this.runsService.deleteStoredMapRuns(mapID);
@@ -629,232 +619,6 @@ export class MapsService {
 
   //#endregion
 
-  //#region Credits
-
-  async getCredits(mapID: number, expand: string[]): Promise<MapCreditDto[]> {
-    if (!(await this.db.map.exists({ where: { id: mapID } })))
-      throw new NotFoundException('Map not found');
-
-    const dbResponse = await this.db.mapCredit.findMany({
-      where: { mapID },
-      include: expandToPrismaIncludes(
-        expand?.filter((x) => ['user'].includes(x))
-      )
-    });
-
-    return dbResponse.map((x) => DtoFactory(MapCreditDto, x));
-  }
-
-  async createCredit(
-    mapID: number,
-    createMapCredit: CreateMapCreditDto,
-    userID: number
-  ): Promise<MapCreditDto> {
-    const map = await this.db.map.findUnique({
-      where: { id: mapID },
-      include: { credits: true }
-    });
-
-    if (!map) throw new NotFoundException('Map not found');
-
-    await this.checkCreditChangePermissions(map, userID);
-
-    if (
-      await this.db.mapCredit.exists({
-        where: {
-          mapID,
-          userID: createMapCredit.userID,
-          type: createMapCredit.type
-        }
-      })
-    )
-      throw new ConflictException('Map credit already exists');
-
-    if (
-      !(await this.db.user.exists({
-        where: { id: createMapCredit.userID }
-      }))
-    )
-      throw new BadRequestException('Credited user does not exist');
-
-    const dbResponse = await this.db.mapCredit.create({
-      data: {
-        type: createMapCredit.type,
-        map: { connect: { id: mapID } },
-        user: { connect: { id: createMapCredit.userID } }
-      }
-    });
-
-    await this.updateMapCreditActivities(dbResponse);
-
-    return DtoFactory(MapCreditDto, dbResponse);
-  }
-
-  async getCredit(
-    mapCreditID: number,
-    expand: string[]
-  ): Promise<MapCreditDto> {
-    const dbResponse = await this.db.mapCredit.findUnique({
-      where: { id: mapCreditID },
-      include: expandToPrismaIncludes(expand)
-    });
-
-    if (!dbResponse) throw new NotFoundException('Map credit not found');
-
-    return DtoFactory(MapCreditDto, dbResponse);
-  }
-
-  async updateCredit(
-    mapCreditID: number,
-    creditUpdate: UpdateMapCreditDto,
-    userID: number
-  ): Promise<void> {
-    if (!creditUpdate.userID && !creditUpdate.type)
-      throw new BadRequestException('No update data provided');
-
-    const credit = await this.db.mapCredit.findUnique({
-      where: { id: mapCreditID },
-      include: { user: true, map: true }
-    });
-
-    if (!credit || !credit.map || !credit.user)
-      throw new NotFoundException('Invalid map credit');
-
-    await this.checkCreditChangePermissions(credit.map, userID);
-
-    if (
-      creditUpdate.userID &&
-      !(await this.db.user.exists({
-        where: { id: creditUpdate.userID }
-      }))
-    )
-      throw new BadRequestException('Credited user does not exist');
-
-    // Check for different credits with same map, user and type, throw if exists
-    if (
-      await this.db.mapCredit.exists({
-        where: {
-          NOT: { id: credit.id },
-          userID: creditUpdate.userID ?? credit.userID,
-          type: creditUpdate.type ?? credit.type,
-          mapID: credit.mapID
-        }
-      })
-    )
-      throw new ConflictException('Cannot have duplicate map credits');
-
-    const data: Prisma.MapCreditUpdateInput = {};
-    if (creditUpdate.userID)
-      data.user = { connect: { id: creditUpdate.userID } };
-    if (creditUpdate.type) data.type = creditUpdate.type;
-
-    const newCredit = await this.db.mapCredit.update({
-      where: { id: mapCreditID },
-      data
-    });
-
-    await this.updateMapCreditActivities(newCredit, credit);
-  }
-
-  async deleteCredit(mapCreditID: number, userID: number): Promise<void> {
-    const credit = await this.db.mapCredit.findUnique({
-      where: { id: mapCreditID },
-      include: {
-        user: true,
-        map: true
-      }
-    });
-
-    if (!credit || !credit.map || !credit.user)
-      throw new NotFoundException('Invalid map credit');
-
-    await this.checkCreditChangePermissions(credit.map, userID);
-
-    await this.db.mapCredit.delete({ where: { id: mapCreditID } });
-
-    await this.updateMapCreditActivities(null, credit);
-  }
-
-  private async checkCreditChangePermissions(map: MapDB, userID: number) {
-    const user = await this.db.user.findUnique({
-      where: { id: userID },
-      select: { roles: true }
-    });
-
-    // Let admins/mods update in any state
-    if (
-      !(
-        Bitflags.has(user.roles, Role.MODERATOR) ||
-        Bitflags.has(user.roles, Role.ADMIN)
-      )
-    ) {
-      if (map.submitterID !== userID)
-        throw new ForbiddenException('User is not the submitter of this map');
-      if (map.status !== MapStatus.NEEDS_REVISION)
-        throw new ForbiddenException('Map is not in NEEDS_REVISION state');
-    }
-  }
-
-  private async updateMapCreditActivities(
-    newCredit?: MapCredit,
-    oldCredit?: MapCredit
-  ): Promise<void> {
-    const deleteOldActivity = () =>
-      this.db.activity.deleteMany({
-        where: {
-          type: ActivityType.MAP_UPLOADED,
-          data: oldCredit.mapID,
-          userID: oldCredit.userID
-        }
-      });
-
-    const createNewActivity = () =>
-      this.db.activity.create({
-        data: {
-          type: ActivityType.MAP_UPLOADED,
-          data: newCredit.mapID,
-          userID: newCredit.userID
-        }
-      });
-
-    // If oldCredit is null, a credit was created
-    if (!oldCredit) {
-      if (newCredit?.type === MapCreditType.AUTHOR) await createNewActivity();
-      return;
-    }
-
-    // If newCredit is null, a credit was deleted
-    if (!newCredit) {
-      if (oldCredit?.type === MapCreditType.AUTHOR) await deleteOldActivity();
-      return;
-    }
-
-    const oldCreditIsAuthor = oldCredit.type === MapCreditType.AUTHOR;
-    const newCreditIsAuthor = newCredit.type === MapCreditType.AUTHOR;
-    const userChanged = oldCredit.userID !== newCredit.userID;
-
-    // If the new credit type was changed to author
-    if (!oldCreditIsAuthor && newCreditIsAuthor) {
-      // Create activity for newCredit.userID
-      await createNewActivity();
-      return;
-    } else if (oldCreditIsAuthor && !newCreditIsAuthor) {
-      // If the new credit type was changed from author to something else
-      // Delete activity for oldCredit.userID
-      await deleteOldActivity();
-      return;
-    } else if (oldCreditIsAuthor && newCreditIsAuthor && userChanged) {
-      // If the credit is still an author but the user changed
-      // Delete activity for oldCredit.userID and create activity for
-      // newCredit.userID
-      await deleteOldActivity();
-      await createNewActivity();
-      return;
-    } else return; // All other cases result in no change in authors
-  }
-
-  //#endregion
-
   //#region Info
 
   async getInfo(mapID: number): Promise<MapInfoDto> {
@@ -893,186 +657,6 @@ export class MapsService {
     await this.db.mapInfo.update({ where: { mapID }, data });
   }
 
-  //#endregion
-
-  //#region Images
-
-  async getImages(mapID: number): Promise<MapImageDto[]> {
-    if (!(await this.db.map.exists({ where: { id: mapID } })))
-      throw new NotFoundException('Map not found');
-
-    const images = await this.db.mapImage.findMany({ where: { mapID } });
-
-    return images.map((x) => DtoFactory(MapImageDto, x));
-  }
-
-  async getImage(imgID: number): Promise<MapImageDto> {
-    const img = await this.db.mapImage.findUnique({ where: { id: imgID } });
-
-    if (!img) throw new NotFoundException('Map image not found');
-
-    return DtoFactory(MapImageDto, img);
-  }
-
-  async createImage(
-    userID: number,
-    mapID: number,
-    imgBuffer: Buffer
-  ): Promise<MapImageDto> {
-    const map = await this.db.map.findUnique({ where: { id: mapID } });
-
-    if (!map) throw new NotFoundException('Map not found');
-
-    if (map.submitterID !== userID)
-      throw new ForbiddenException('User is not the submitter of the map');
-
-    if (map.status !== MapStatus.NEEDS_REVISION)
-      throw new ForbiddenException('Map is not in NEEDS_REVISION state');
-
-    const images = await this.db.mapImage.findMany({ where: { mapID } });
-    let imageCount = images.length;
-    if (map.thumbnailID) imageCount--; // Don't count the thumbnail towards this limit
-    if (imageCount >= this.config.get('limits.mapImageUploads'))
-      throw new ConflictException('Map image file limit reached');
-
-    // It may seem strange to create an entry with nothing but a key into Map,
-    // but we're doing it so we get an ID from Postgres, and can store the image
-    // file at a corresponding URL.
-    const newImage = await this.db.mapImage.create({ data: { mapID } });
-
-    const uploadedImages = await this.storeMapImage(imgBuffer, newImage.id);
-
-    if (!uploadedImages) {
-      await this.db.mapImage.delete({ where: { id: newImage.id } });
-      throw new BadGatewayException('Error uploading image to CDN');
-    }
-
-    return DtoFactory(MapImageDto, newImage);
-  }
-
-  async updateImage(
-    userID: number,
-    imgID: number,
-    imgBuffer: Buffer
-  ): Promise<void> {
-    await this.editMapImageChecks(userID, imgID);
-
-    const uploadedImages = await this.storeMapImage(imgBuffer, imgID);
-
-    if (!uploadedImages)
-      throw new BadGatewayException('Failed to upload image to CDN');
-  }
-
-  async deleteImage(userID: number, imgID: number): Promise<void> {
-    await this.editMapImageChecks(userID, imgID);
-
-    await Promise.all([
-      this.deleteStoredMapImage(imgID),
-      this.db.mapImage.delete({ where: { id: imgID } })
-    ]);
-  }
-
-  private async editMapImageChecks(
-    userID: number,
-    imgID: number
-  ): Promise<void> {
-    const image = await this.db.mapImage.findUnique({ where: { id: imgID } });
-
-    if (!image) throw new NotFoundException('Image not found');
-
-    const map = await this.db.map.findUnique({ where: { id: image.mapID } });
-
-    if (map.submitterID !== userID)
-      throw new ForbiddenException('User is not the submitter of the map');
-
-    if (map.status !== MapStatus.NEEDS_REVISION)
-      throw new ForbiddenException('Map is not in NEEDS_REVISION state');
-  }
-
-  private async editSaveMapImageFile(
-    imgBuffer: Buffer,
-    fileName: string,
-    width: number,
-    height: number
-  ): Promise<FileStoreCloudFile> {
-    try {
-      return this.fileCloudService.storeFileCloud(
-        await sharp(imgBuffer)
-          .resize(width, height, { fit: 'inside' })
-          .jpeg({ mozjpeg: true })
-          .toBuffer(),
-        fileName
-      );
-    } catch {
-      // This looks bad, but sharp is very non-specific about its errors
-      throw new BadRequestException('Invalid image file');
-    }
-  }
-
-  async storeMapImage(
-    imgBuffer: Buffer,
-    imgID: number
-  ): Promise<FileStoreCloudFile[]> {
-    return Promise.all([
-      this.editSaveMapImageFile(imgBuffer, `img/${imgID}-small.jpg`, 480, 360),
-      this.editSaveMapImageFile(
-        imgBuffer,
-        `img/${imgID}-medium.jpg`,
-        1280,
-        720
-      ),
-      this.editSaveMapImageFile(imgBuffer, `img/${imgID}-large.jpg`, 1920, 1080)
-    ]);
-  }
-
-  async deleteStoredMapImage(imgID: number): Promise<void> {
-    await Promise.all([
-      this.fileCloudService.deleteFileCloud(`img/${imgID}-small.jpg`),
-      this.fileCloudService.deleteFileCloud(`img/${imgID}-medium.jpg`),
-      this.fileCloudService.deleteFileCloud(`img/${imgID}-large.jpg`)
-    ]);
-  }
-
-  //#endregion
-
-  //#region Thumbnails
-
-  async updateThumbnail(
-    userID: number,
-    mapID: number,
-    imgBuffer: Buffer
-  ): Promise<void> {
-    let map = await this.db.map.findUnique({
-      where: { id: mapID },
-      select: { thumbnailID: true, submitterID: true, status: true }
-    });
-
-    if (!map) throw new NotFoundException('Map not found');
-    if (map.submitterID !== userID)
-      throw new ForbiddenException('User is not the submitter of the map');
-    if (map.status !== MapStatus.NEEDS_REVISION)
-      throw new ForbiddenException('Map is not in NEEDS_REVISION state');
-
-    if (!map.thumbnailID) {
-      const newThumbnail = await this.db.mapImage.create({ data: { mapID } });
-      map = await this.db.map.update({
-        where: { id: mapID },
-        data: { thumbnail: { connect: { id: newThumbnail.id } } }
-      });
-    }
-
-    const thumbnail = await this.db.mapImage.findUnique({
-      where: { id: map.thumbnailID }
-    });
-
-    const uploadedImages = await this.storeMapImage(imgBuffer, thumbnail.id);
-    if (!uploadedImages) {
-      // If the images failed to upload, we want to delete the map image object
-      // if there was no previous thumbnail
-      await this.db.mapImage.delete({ where: { id: thumbnail.id } });
-      throw new BadGatewayException('Failed to upload image to CDN');
-    }
-  }
   //#endregion
 
   //#region Zones
